test(onboard): split vllm selection suite
